refactor: shell.rsとshell_ui.rsのテストコードを別ファイルに抽出
AST Linterのファイル行数制約に従い、メインファイルの肥大化を解消するため
テストコードのみを src 内の別ファイル (shell_tests.rs, shell_ui_tests.rs) に分離。
#[cfg(test)] include! を用いることで内部スコープへのアクセス権限を維持。
