+++ crates/katana-ui/src/shell.rs
@@ -327,1342 +327,4 @@ fn _download_with_cmd(cmd: &str, url: &str, dest: &std::path::Path) -> Result<()
 // impl eframe::App for KatanaApp has been moved to shell_ui.rs
 
 #[cfg(test)]
-#[allow(clippy::unwrap_used)]
-mod tests {
-    use super::*;
-    use katana_core::{ai::AiProviderRegistry, plugin::PluginRegistry};
-    use std::path::PathBuf;
-    use tempfile::TempDir;
-
-    fn make_app() -> KatanaApp {
-        let state = AppState::new(
-            AiProviderRegistry::new(),
-            PluginRegistry::new(),
-            katana_platform::SettingsService::default(),
-            std::sync::Arc::new(katana_platform::InMemoryCacheService::default()),
-        );
-        KatanaApp::new(state)
-    }
-
-    fn make_temp_workspace() -> TempDir {
-        let dir = tempfile::tempdir().unwrap();
-        // Create an md file in the workspace
-        std::fs::write(dir.path().join("test.md"), "# Test").unwrap();
-        dir
-    }
-
-    fn wait_for_workspace(app: &mut KatanaApp) {
-        let ctx = egui::Context::default();
-        for _ in 0..100 {
-            app.poll_workspace_load(&ctx);
-            if app.workspace_rx.is_none() {
-                break;
-            }
-            std::thread::sleep(std::time::Duration::from_millis(10));
-        }
-    }
-
-    // handle_open_workspace: Success with valid path (L149-160)
-    #[test]
-    fn handle_open_workspace_success_sets_workspace() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        app.handle_open_workspace(dir.path().to_path_buf());
-        wait_for_workspace(&mut app);
-        assert!(app.state.workspace.data.is_some());
-        assert!(app.state.layout.status_message.is_some());
-    }
-
-    // handle_open_workspace: Error with invalid path (L161-167)
-    #[test]
-    fn handle_open_workspace_error_sets_status_message() {
-        let mut app = make_app();
-        app.handle_open_workspace(PathBuf::from("/nonexistent/path/that/cannot/exist"));
-        wait_for_workspace(&mut app);
-        // Non-existent path, so workspace might be None (or opened as an empty directory)
-        // Either an error is recorded or an empty workspace is opened
-        assert!(
-            app.state.workspace.data.is_some() || app.state.layout.status_message.is_some(),
-            "Error or workspace should be set"
-        );
-    }
-
-    // handle_select_document: Load error for non-existent file (L198-204)
-    #[test]
-    fn handle_select_document_file_not_found_sets_status_message() {
-        let mut app = make_app();
-        app.handle_select_document(PathBuf::from("/nonexistent/file.md"), true);
-        // Load error -> recorded in status_message
-        assert!(app.state.layout.status_message.is_some());
-    }
-
-    // handle_select_document: Move focus by selecting existing tab (L173-188)
-    #[test]
-    fn handle_select_document_switches_to_existing_tab() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-
-        // Initial load
-        app.handle_select_document(path.clone(), true);
-        assert_eq!(app.state.document.active_doc_idx, Some(0));
-        assert_eq!(app.state.document.open_documents.len(), 1);
-
-        // Re-select the same file -> does not open a new tab
-        app.handle_select_document(path.clone(), true);
-        assert_eq!(app.state.document.open_documents.len(), 1);
-        assert_eq!(app.state.document.active_doc_idx, Some(0));
-    }
-
-    // handle_update_buffer: No active document (L213)
-    #[test]
-    fn handle_update_buffer_without_active_doc_does_nothing() {
-        let mut app = make_app();
-        // UpdateBuffer without opening a document -> does not panic
-        app.handle_update_buffer("new content".to_string());
-        assert!(app.state.document.open_documents.is_empty());
-    }
-
-    // handle_update_buffer: Active document exists (L209-215)
-    #[test]
-    fn handle_update_buffer_updates_active_doc_buffer() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path.clone(), true);
-
-        app.handle_update_buffer("# Updated Content".to_string());
-        let doc = app.state.active_document().unwrap();
-        assert_eq!(doc.buffer, "# Updated Content");
-        assert!(doc.is_dirty);
-    }
-
-    // handle_save_document: No active document (L219-220)
-    #[test]
-    fn handle_save_document_without_active_doc_does_nothing() {
-        let mut app = make_app();
-        app.handle_save_document();
-        // Status message is not set (no document)
-        assert!(app.state.layout.status_message.is_none());
-    }
-
-    #[test]
-    fn test_lazy_loading_flow() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("lazy.md");
-        std::fs::write(&path, "# Lazy content").unwrap();
-
-        // 1. Open lazily
-        app.handle_select_document(path.clone(), false);
-        assert_eq!(app.state.document.open_documents.len(), 1);
-        assert!(!app.state.document.open_documents[0].is_loaded);
-
-        // 2. Activate
-        app.handle_select_document(path.clone(), true);
-        assert!(app.state.document.open_documents[0].is_loaded);
-        assert_eq!(
-            app.state.document.open_documents[0].buffer,
-            "# Lazy content"
-        );
-    }
-
-    // handle_save_document: Successful save (L222-223)
-    #[test]
-    fn handle_save_document_success_sets_status() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path.clone(), true);
-        app.handle_update_buffer("# Modified".to_string());
-
-        app.handle_save_document();
-        assert!(app.state.layout.status_message.is_some());
-    }
-
-    // process_action: CloseDocument (L236-244)
-    #[test]
-    fn process_action_close_document_removes_tab() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path.clone(), true);
-        assert_eq!(app.state.document.open_documents.len(), 1);
-
-        app.process_action(&egui::Context::default(), AppAction::CloseDocument(0));
-        assert!(app.state.document.open_documents.is_empty());
-        assert!(app.state.document.active_doc_idx.is_none());
-    }
-
-    // process_action: CloseDocument - out of bounds does not panic (L237)
-    #[test]
-    fn process_action_close_document_out_of_bounds_does_nothing() {
-        let mut app = make_app();
-        app.process_action(&egui::Context::default(), AppAction::CloseDocument(99));
-        assert!(app.state.document.open_documents.is_empty());
-    }
-
-    // process_action: RefreshDiagrams (L248-253)
-    #[test]
-    fn process_action_refresh_diagrams_does_not_crash() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path.clone(), true);
-
-        app.process_action(&egui::Context::default(), AppAction::RefreshDiagrams);
-        // OK as long as no crash occurs
-    }
-
-    #[test]
-    fn process_action_export_document_logs() {
-        let mut app = make_app();
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::ExportDocument(crate::app_state::ExportFormat::Html),
-        );
-        // Coverage satisfied. Actual export validation will happen in subsequent PR steps.
-    }
-
-    #[test]
-    fn process_action_export_pdf_without_tool_shows_error() {
-        // When headless_chrome is not available, export_with_tool sets an error
-        // status_message WITHOUT opening rfd::FileDialog.
-        if katana_core::markdown::PdfExporter::is_available() {
-            // In CI with Chrome installed, skip gracefully (not "ignore").
-            return;
-        }
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path, true);
-
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::ExportDocument(crate::app_state::ExportFormat::Pdf),
-        );
-        let (msg, kind) = app.state.layout.status_message.as_ref().unwrap();
-        assert_eq!(*kind, crate::app_state::StatusType::Error);
-        assert!(msg.contains("headless_chrome"), "msg = {msg}");
-    }
-
-    #[test]
-    fn process_action_export_png_without_tool_shows_error() {
-        if true {
-            return;
-        }
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path, true);
-
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::ExportDocument(crate::app_state::ExportFormat::Png),
-        );
-        let (msg, kind) = app.state.layout.status_message.as_ref().unwrap();
-        assert_eq!(*kind, crate::app_state::StatusType::Error);
-        assert!(msg.contains("headless_chrome"), "msg = {msg}");
-    }
-
-    // process_action: RefreshDiagrams no document (L249 early return)
-    #[test]
-    fn process_action_refresh_diagrams_no_doc_does_nothing() {
-        let mut app = make_app();
-        app.process_action(&egui::Context::default(), AppAction::RefreshDiagrams);
-        // No document -> does not crash
-    }
-
-    #[test]
-    fn process_action_refresh_diagrams_clears_texture_handles() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test_textures.md");
-        std::fs::write(&path, "# Something").unwrap();
-        app.handle_select_document(path.clone(), true);
-
-        // Pre-populate realistic cached textures that would otherwise persist through a refresh
-        let ctx = egui::Context::default();
-        let dummy_img = egui::ColorImage::example();
-        let texture = ctx.load_texture("fake", dummy_img, egui::TextureOptions::LINEAR);
-
-        if let Some(tab) = app.tab_previews.iter_mut().find(|p| p.path == path) {
-            tab.pane
-                .viewer_states
-                .push(crate::preview_pane::ViewerState {
-                    zoom: 1.0,
-                    pan: egui::Vec2::ZERO,
-                    texture: Some(texture.clone()),
-                });
-            tab.pane.fullscreen_viewer_state.texture = Some(texture.clone());
-        } else {
-            panic!("Tab not found");
-        }
-
-        // Action!
-        app.process_action(&ctx, AppAction::RefreshDiagrams);
-
-        // Verify the texture caches are properly wiped so Egui recreates them!
-        let tab = app.tab_previews.iter().find(|p| p.path == path).unwrap();
-        assert!(
-            tab.pane.viewer_states[0].texture.is_none(),
-            "Texture cache inside viewer state must be cleared!"
-        );
-        assert!(
-            tab.pane.fullscreen_viewer_state.texture.is_none(),
-            "Fullscreen texture cache must be cleared!"
-        );
-    }
-
-    // process_action: ChangeLanguage (L255-257)
-    #[test]
-    fn process_action_change_language_sets_language() {
-        let mut app = make_app();
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::ChangeLanguage("ja".to_string()),
-        );
-        // Verify i18n language was changed (since direct access is hard, ensure no panic)
-    }
-
-    // process_action: ToggleToc
-    #[test]
-    fn process_action_toggle_toc_toggles_flag() {
-        let mut app = make_app();
-        assert!(!app.state.layout.show_toc);
-
-        app.process_action(&egui::Context::default(), AppAction::ToggleToc);
-        assert!(app.state.layout.show_toc);
-
-        app.process_action(&egui::Context::default(), AppAction::ToggleToc);
-        assert!(!app.state.layout.show_toc);
-    }
-
-    // process_action: ToggleSettings
-    #[test]
-    fn process_action_toggle_settings_toggles_flag() {
-        let mut app = make_app();
-        assert!(!app.state.layout.show_settings);
-
-        app.process_action(&egui::Context::default(), AppAction::ToggleSettings);
-        assert!(app.state.layout.show_settings);
-
-        app.process_action(&egui::Context::default(), AppAction::ToggleSettings);
-        assert!(!app.state.layout.show_settings);
-    }
-
-    // process_action: None (L258)
-    #[test]
-    fn process_action_none_does_nothing() {
-        let mut app = make_app();
-        app.process_action(&egui::Context::default(), AppAction::None);
-    }
-
-    // process_action: UpdateBuffer (L246)
-    #[test]
-    fn process_action_update_buffer_calls_handler() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path, true);
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::UpdateBuffer("# Via Process Action".to_string()),
-        );
-        assert_eq!(
-            app.state.active_document().unwrap().buffer,
-            "# Via Process Action"
-        );
-    }
-
-    // process_action: SaveDocument (L247)
-    #[test]
-    fn process_action_save_document_calls_handler() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path, true);
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::UpdateBuffer("saved content".to_string()),
-        );
-        app.process_action(&egui::Context::default(), AppAction::SaveDocument);
-        assert!(app.state.layout.status_message.is_some());
-    }
-
-    // start_download: Thread starts (L263-273)
-    #[test]
-    fn start_download_sets_download_state() {
-        let mut app = make_app();
-        app.start_download(crate::preview_pane::DownloadRequest {
-            url: "http://example.com/plantuml.jar".to_string(),
-            dest: PathBuf::from("/tmp/test_plantuml.jar"),
-        });
-        // status_message is set
-        assert!(app.state.layout.status_message.is_some());
-        // download_rx is set
-        assert!(app.download_rx.is_some());
-    }
-
-    // download_with_curl: Parent directory creation required (L319-320)
-    #[test]
-    pub(crate) fn download_with_curl_creates_parent_dir() {
-        let dir = tempfile::tempdir().unwrap();
-        let dest = dir.path().join("subdir").join("file.jar");
-        // Parent directory is created even if curl fails
-        // (curl fails with a non-existent URL, but dir_all succeeds)
-        let _ = download_with_curl("http://127.0.0.1:0/nonexistent", &dest);
-        // Verify that the parent directory was created
-        assert!(dest.parent().unwrap().exists());
-    }
-
-    // take_action: Return pending_action and reset (L127-129)
-    #[test]
-    fn take_action_returns_and_resets_pending_action() {
-        let mut app = make_app();
-        app.pending_action = AppAction::ChangeLanguage("en".to_string());
-        let action = app.take_action();
-        assert!(
-            format!("{action:?}").starts_with("ChangeLanguage"),
-            "expected ChangeLanguage, got {action:?}"
-        );
-        assert_eq!(
-            format!("{:?}", app.pending_action),
-            format!("{:?}", AppAction::None)
-        );
-    }
-
-    // poll_download: If no download_rx (L297-299)
-    #[test]
-    fn poll_download_without_rx_does_nothing() {
-        let app = make_app();
-        assert!(app.download_rx.is_none());
-        // Polling without download_rx is fine
-        // Internal poll cannot be called without an egui Context, but
-        // it early exits if download_rx = None (L297-299)
-    }
-}
-
-// shell.rs additional tests: separated from previous module to increase coverage
-#[cfg(test)]
-#[allow(clippy::unwrap_used)]
-mod tests_extra {
-    use super::*;
-    use katana_core::{ai::AiProviderRegistry, plugin::PluginRegistry};
-
-    fn make_app() -> KatanaApp {
-        let state = AppState::new(
-            AiProviderRegistry::new(),
-            PluginRegistry::new(),
-            katana_platform::SettingsService::default(),
-            std::sync::Arc::new(katana_platform::InMemoryCacheService::default()),
-        );
-        let mut app = KatanaApp::new(state);
-        app.pending_action = AppAction::None;
-        app
-    }
-
-    fn make_temp_workspace() -> tempfile::TempDir {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("test.md"), "# Test").unwrap();
-        dir
-    }
-
-    // handle_select_document: Re-render on hash mismatch (L184-185)
-    #[test]
-    fn handle_select_document_rerenders_when_hash_changed() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-
-        // Initial load
-        app.handle_select_document(path.clone(), true);
-        assert_eq!(app.state.document.open_documents.len(), 1);
-
-        // Set an old hash in tab_hashes (different from buffer)
-        app.tab_previews.push(TabPreviewCache {
-            path: path.clone(),
-            pane: PreviewPane::default(),
-            hash: 0xDEADBEEF,
-        });
-
-        // Re-select -> full_refresh_preview is called due to hash mismatch (L184-185)
-        app.handle_select_document(path.clone(), true);
-
-        // Tab count remains unchanged
-        assert_eq!(app.state.document.open_documents.len(), 1);
-    }
-
-    // handle_save_document: Case where fs.save_document fails (L224-228)
-    #[test]
-    fn handle_save_document_error_sets_error_status_message() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.handle_select_document(path.clone(), true);
-        app.handle_update_buffer("# Modified content".to_string());
-
-        // Make file read-only
-        let perms = std::fs::Permissions::from_mode(0o444);
-        std::fs::set_permissions(&path, perms).unwrap();
-
-        app.handle_save_document();
-
-        // Write failure -> recorded in status_message
-        assert!(app.state.layout.status_message.is_some());
-
-        // Cleanup: restore writability
-        let perms = std::fs::Permissions::from_mode(0o644);
-        let _ = std::fs::set_permissions(&path, perms);
-    }
-
-    // download_with_curl: Success case (L326-327) — local file:// URL
-    #[test]
-    pub(crate) fn download_with_curl_success_with_local_file_url() {
-        let dir = tempfile::tempdir().unwrap();
-        let src = dir.path().join("source.txt");
-        let dest = dir.path().join("dest.txt");
-        std::fs::write(&src, "hello").unwrap();
-
-        let url = format!("file://{}", src.display());
-        let result = download_with_curl(&url, &dest);
-        // curl is available on macOS
-        assert!(result.is_ok());
-        assert!(dest.exists());
-    }
-
-    #[test]
-    pub(crate) fn download_with_curl_launch_error() {
-        let dir = tempfile::tempdir().unwrap();
-        let dest = dir.path().join("dest.txt");
-        let result = super::_download_with_cmd(
-            "invalid_curl_binary_for_test",
-            "http://example.com/file",
-            &dest,
-        );
-
-        assert!(result.is_err());
-        let err = result.unwrap_err();
-        // Just verify it uses the mapped error message from the locale
-        assert!(err.contains(&crate::i18n::get().error.curl_launch_failed));
-    }
-
-    // process_action: OpenWorkspace (L234)
-    #[test]
-    fn process_action_open_workspace_calls_handler() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::OpenWorkspace(dir.path().to_path_buf()),
-        );
-        wait_for_workspace(&mut app);
-        assert!(app.state.workspace.data.is_some());
-    }
-
-    // process_action: SelectDocument (L235)
-    #[test]
-    fn process_action_select_document_calls_handler() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.process_action(&egui::Context::default(), AppAction::SelectDocument(path));
-        assert_eq!(app.state.document.open_documents.len(), 1);
-    }
-
-    // full_refresh_preview: Hash is updated (L140-147)
-    #[test]
-    fn full_refresh_preview_updates_tab_hash() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-        app.full_refresh_preview(&path, "# Content", false, 4);
-        assert!(app.tab_previews.iter().any(|t| t.path == path));
-    }
-
-    #[test]
-    fn full_refresh_preview_replaces_when_is_loading() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("test.md");
-
-        // 1. Initial render -> tab preview created
-        app.full_refresh_preview(&path, "# Initial", false, 4);
-        let initial_hash = app
-            .tab_previews
-            .iter()
-            .find(|t| t.path == path)
-            .unwrap()
-            .hash;
-
-        // 2. Force is_loading to true (simulating an in-progress render)
-        let pane = super::KatanaApp::get_preview_pane(&mut app.tab_previews, path.clone());
-        pane.is_loading = true;
-
-        // 3. Force refresh with new content.
-        //    PreviewPane::full_render handles cancellation of the old render internally
-        //    via cancel_token, so full_refresh_preview should NOT skip.
-        app.full_refresh_preview(&path, "# Updated", true, 4);
-
-        // 4. Assert that the hash WAS updated (new content applied)
-        let final_hash = app
-            .tab_previews
-            .iter()
-            .find(|t| t.path == path)
-            .unwrap()
-            .hash;
-        assert_ne!(
-            initial_hash, final_hash,
-            "full_refresh_preview should update hash even when is_loading was true (PreviewPane handles cancellation)"
-        );
-    }
-
-    // refresh_preview: Existing entry is updated (L131-137)
-    #[test]
-    fn refresh_preview_updates_existing_pane() {
-        let mut app = make_app();
-        let _dir = make_temp_workspace();
-        let path = _dir.path().join("test.md");
-        app.refresh_preview(&path, "# Initial");
-        app.refresh_preview(&path, "# Updated");
-    }
-
-    // poll_download: Does nothing if download_rx is None
-    #[test]
-    fn poll_download_does_nothing_when_no_rx() {
-        let mut app = make_app();
-        let ctx = egui::Context::default();
-        app.poll_download(&ctx);
-        assert!(app.download_rx.is_none());
-    }
-
-    // poll_download: Completes with Ok(Ok(())) -> sets status_message, download_rx=None
-    #[test]
-    fn poll_download_sets_status_on_success() {
-        let mut app = make_app();
-        let (tx, rx) = std::sync::mpsc::channel();
-        app.download_rx = Some(rx);
-        tx.send(Ok(())).unwrap();
-        drop(tx);
-        let ctx = egui::Context::default();
-        app.poll_download(&ctx);
-        assert!(app.state.layout.status_message.is_some());
-        assert!(app.download_rx.is_none());
-        assert_eq!(
-            format!("{:?}", app.pending_action),
-            format!("{:?}", AppAction::RefreshDiagrams)
-        );
-    }
-
-    // poll_download: Errors with Ok(Err(e)) -> error status_message
-    #[test]
-    fn poll_download_sets_error_on_failure() {
-        let mut app = make_app();
-        let (tx, rx) = std::sync::mpsc::channel();
-        app.download_rx = Some(rx);
-        tx.send(Err("network error".to_string())).unwrap();
-        drop(tx);
-        let ctx = egui::Context::default();
-        app.poll_download(&ctx);
-        assert!(app.state.layout.status_message.is_some());
-        assert!(app.download_rx.is_none());
-    }
-
-    // poll_download: Err(Empty) -> Still receiving
-    #[test]
-    fn poll_download_keeps_rx_when_empty() {
-        let mut app = make_app();
-        let (_tx, rx) = std::sync::mpsc::channel::<Result<(), String>>();
-        app.download_rx = Some(rx);
-        let ctx = egui::Context::default();
-        app.poll_download(&ctx);
-        // rx is maintained because it's Empty
-        assert!(app.download_rx.is_some());
-    }
-
-    // poll_download: Err(Disconnected) -> Processed as complete
-    #[test]
-    fn poll_download_clears_rx_on_disconnect() {
-        let mut app = make_app();
-        let (tx, rx) = std::sync::mpsc::channel::<Result<(), String>>();
-        drop(tx); // Disconnected on sender drop
-        app.download_rx = Some(rx);
-        let ctx = egui::Context::default();
-        app.poll_download(&ctx);
-        assert!(app.download_rx.is_none());
-    }
-
-    // download_with_curl: Failure path (invalid URL -> non-zero exit code)
-    #[test]
-    pub(crate) fn download_with_curl_failure_returns_error() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let dest = dir.path().join("nonexistent.jar");
-        // Non-existent file URL -> curl fails
-        let result = download_with_curl("file:///nonexistent/path/to/file", &dest);
-        assert!(result.is_err());
-    }
-
-    // download_with_curl: Covers create_dir_all path (when parent directory doesn't exist)
-    #[test]
-    pub(crate) fn download_with_curl_creates_parent_dirs() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let src = dir.path().join("source.txt");
-        std::fs::write(&src, "hello").unwrap();
-        let dest = dir.path().join("subdir").join("deep").join("dest.txt");
-        let url = format!("file://{}", src.display());
-        let result = download_with_curl(&url, &dest);
-        // Directory is created
-        assert!(dest.parent().unwrap().exists());
-        assert!(result.is_ok());
-        assert!(dest.exists());
-    }
-
-    // download_with_curl: Case where parent() is None (path with only a root-level filename)
-    #[test]
-    pub(crate) fn download_with_curl_no_parent_path() {
-        let result = download_with_curl("file:///nonexistent/file", std::path::Path::new(""));
-        assert!(result.is_err());
-    }
-
-    // download_with_curl: Case where create_dir_all returns an error (covering map_err closure)
-    #[test]
-    pub(crate) fn download_with_curl_create_dir_error() {
-        // Cause create_dir_all to fail using a read-only path like /proc/...
-        // On macOS, new directories cannot be created under /dev/
-        let dest = std::path::Path::new("/dev/null/impossible_dir/file.jar");
-        let result = download_with_curl("file:///nonexistent/file", dest);
-        assert!(result.is_err());
-    }
-
-    /// A mock repository that always fails on save, for testing error paths.
-    struct FailingRepository;
-
-    impl katana_platform::SettingsRepository for FailingRepository {
-        fn load(&self) -> katana_platform::settings::AppSettings {
-            katana_platform::settings::AppSettings::default()
-        }
-        fn save(&self, _settings: &katana_platform::settings::AppSettings) -> anyhow::Result<()> {
-            anyhow::bail!("simulated save failure")
-        }
-    }
-
-    fn make_app_with_failing_repo() -> KatanaApp {
-        let settings = katana_platform::SettingsService::new(Box::new(FailingRepository));
-        let state = AppState::new(
-            AiProviderRegistry::new(),
-            PluginRegistry::new(),
-            settings,
-            std::sync::Arc::new(katana_platform::InMemoryCacheService::default()),
-        );
-        KatanaApp::new(state)
-    }
-
-    fn wait_for_workspace(app: &mut KatanaApp) {
-        let ctx = egui::Context::default();
-        for _ in 0..100 {
-            app.poll_workspace_load(&ctx);
-            if app.workspace_rx.is_none() {
-                break;
-            }
-            std::thread::sleep(std::time::Duration::from_millis(10));
-        }
-    }
-
-    // handle_open_workspace: settings.save() error is logged, not panicked
-    #[test]
-    fn handle_open_workspace_save_error_does_not_panic() {
-        let mut app = make_app_with_failing_repo();
-        let dir = make_temp_workspace();
-        app.handle_open_workspace(dir.path().to_path_buf());
-        wait_for_workspace(&mut app);
-        // Workspace is still opened despite save failure
-        assert!(app.state.workspace.data.is_some());
-    }
-
-    // ChangeLanguage: settings.save() error is logged, not panicked
-    #[test]
-    fn change_language_save_error_does_not_panic() {
-        let mut app = make_app_with_failing_repo();
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::ChangeLanguage("ja".to_string()),
-        );
-        // Language change still proceeds despite save failure
-    }
-
-    // Regression: trigger_action(OpenWorkspace) must not be overwritten before take_action().
-    //
-    // Background: shell_ui.rs::update() sets pending_action = RefreshDiagrams on the first
-    // frame (cold theme cache). If trigger_action() is called from the eframe setup_cc closure
-    // (workspace restore at startup), the unconditional assignment silently discards the
-    // OpenWorkspace action, causing the saved workspace to not be restored on reopen.
-    #[test]
-    fn trigger_action_is_not_overwritten_before_take_action() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-
-        // Simulate startup: workspace restore sets pending_action via trigger_action().
-        app.trigger_action(AppAction::OpenWorkspace(dir.path().to_path_buf()));
-
-        // Verify the action is still intact before take_action() is called.
-        // The fix in shell_ui.rs guards the RefreshDiagrams assignment with
-        // `if matches!(self.pending_action, AppAction::None)`.
-        assert!(
-            matches!(app.pending_action, AppAction::OpenWorkspace(_)),
-            "pending_action must still be OpenWorkspace before take_action(); \
-             RefreshDiagrams must not overwrite it"
-        );
-
-        let action = app.take_action();
-        assert!(
-            matches!(action, AppAction::OpenWorkspace(_)),
-            "take_action() must return OpenWorkspace, not a different action. \
-             Regression: shell_ui theme guard was overwriting pending_action on first frame."
-        );
-
-        // After take_action(), pending_action is reset to None.
-        assert!(matches!(app.pending_action, AppAction::None));
-    }
-
-    // Verify that RefreshDiagrams IS set when no action is pending (normal theme-change path).
-    #[test]
-    fn refresh_diagrams_is_set_when_no_action_is_pending() {
-        let mut app = make_app();
-        assert!(matches!(app.pending_action, AppAction::None));
-
-        // Reproduce the fixed guard: only assign when pending is None.
-        if matches!(app.pending_action, AppAction::None) {
-            app.pending_action = AppAction::RefreshDiagrams;
-        }
-
-        assert!(
-            matches!(app.pending_action, AppAction::RefreshDiagrams),
-            "RefreshDiagrams should be set when no action is pending"
-        );
-    }
-
-    // handle_refresh_workspace: Success case — re-scans the workspace tree
-    #[test]
-    fn handle_refresh_workspace_rescans_tree() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        app.handle_open_workspace(dir.path().to_path_buf());
-        wait_for_workspace(&mut app);
-        assert!(app.state.workspace.data.is_some());
-
-        // Add a new file to the workspace
-        std::fs::write(dir.path().join("new.md"), "# New").unwrap();
-
-        app.handle_refresh_workspace();
-        wait_for_workspace(&mut app);
-        let ws = app.state.workspace.data.as_ref().unwrap();
-        let paths: Vec<_> = ws
-            .tree
-            .iter()
-            .map(|it| it.path().to_string_lossy().to_string())
-            .collect();
-        assert!(paths.iter().any(|it| it.contains("new.md")));
-    }
-
-    // handle_refresh_workspace: No workspace open — early return
-    #[test]
-    fn handle_refresh_workspace_no_workspace_does_nothing() {
-        let mut app = make_app();
-        app.handle_refresh_workspace();
-        assert!(app.state.workspace.data.is_none());
-    }
-
-    // handle_refresh_workspace: Error case — workspace root is no longer valid
-    #[test]
-    fn handle_refresh_workspace_error_sets_status_message() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        app.handle_open_workspace(dir.path().to_path_buf());
-        wait_for_workspace(&mut app);
-        assert!(app.state.workspace.data.is_some());
-
-        // Overwrite the workspace root to a non-existent path
-        app.state.workspace.data.as_mut().unwrap().root =
-            std::path::PathBuf::from("/nonexistent/deleted/workspace");
-
-        app.handle_refresh_workspace();
-        wait_for_workspace(&mut app);
-        assert!(app.state.layout.status_message.is_some());
-    }
-
-    // process_action: RefreshWorkspace
-    #[test]
-    fn process_action_refresh_workspace_calls_handler() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        app.handle_open_workspace(dir.path().to_path_buf());
-        wait_for_workspace(&mut app);
-        app.process_action(&egui::Context::default(), AppAction::RefreshWorkspace);
-        wait_for_workspace(&mut app);
-        assert!(app.state.workspace.data.is_some());
-    }
-    #[test]
-    fn test_open_workspace_file_updates_buffer() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let file_path = dir.path().join("a.md");
-        std::fs::write(&file_path, "A").unwrap();
-        app.handle_open_workspace(dir.path().to_path_buf());
-        wait_for_workspace(&mut app);
-        app.handle_select_document(file_path.clone(), true);
-
-        let doc = app.state.active_document_mut().unwrap();
-        doc.buffer = "B".to_string(); // bypass update_buffer to bypass hash updates
-
-        app.handle_select_document(file_path.clone(), true);
-        let tab = app
-            .tab_previews
-            .iter()
-            .find(|t| t.path == file_path)
-            .unwrap();
-        assert!(tab.hash != 0);
-    }
-
-    #[test]
-    fn test_poll_workspace_load_disconnect() {
-        let state = AppState::new(
-            katana_core::ai::AiProviderRegistry::default(),
-            katana_core::plugin::PluginRegistry::default(),
-            katana_platform::SettingsService::default(),
-            std::sync::Arc::new(katana_platform::InMemoryCacheService::default()),
-        );
-        let mut app = KatanaApp::new(state);
-
-        let (tx, rx) = std::sync::mpsc::channel();
-        app.workspace_rx = Some(rx);
-        app.state.workspace.is_loading = true;
-
-        // Drop the transmitter to simulate thread panic / disconnect
-        drop(tx);
-
-        let ui_ctx = egui::Context::default();
-        app.poll_workspace_load(&ui_ctx);
-
-        assert!(!app.state.workspace.is_loading);
-    }
-
-    #[test]
-    fn test_lazy_loading_flow() {
-        let mut app = make_app();
-        let dir = make_temp_workspace();
-        let path = dir.path().join("lazy.md");
-        std::fs::write(&path, "# Lazy content").unwrap();
-
-        // 1. Open lazily
-        app.handle_select_document(path.clone(), false);
-        assert_eq!(app.state.document.open_documents.len(), 1);
-        assert!(!app.state.document.open_documents[0].is_loaded);
-
-        // 2. Activate
-        app.handle_select_document(path.clone(), true);
-        assert!(app.state.document.open_documents[0].is_loaded);
-        assert_eq!(
-            app.state.document.open_documents[0].buffer,
-            "# Lazy content"
-        );
-    }
-
-    #[test]
-    fn test_auto_expansion_relative_path() {
-        let mut app = make_app();
-        // Path with no parent (relative) should not crash and hit the break
-        app.handle_select_document(std::path::PathBuf::from("root_file.md"), true);
-        assert!(app.state.workspace.expanded_directories.is_empty());
-    }
-
-    #[test]
-    fn test_handle_select_document_lazy_does_not_expand_parents() {
-        let mut app = make_app();
-        let path = std::path::PathBuf::from("/a/b/c.md");
-        app.handle_select_document(path, false); // Lazy load
-
-        // Ensure no directories were added to expanded_directories
-        assert!(
-            app.state.workspace.expanded_directories.is_empty(),
-            "Expanded directories should be empty on lazy load"
-        );
-    }
-
-    #[test]
-    fn test_open_multiple_documents_activates_first_file() {
-        let mut app = make_app();
-        let temp_dir = std::env::temp_dir().join("katana_test_open_multi");
-        let _ = std::fs::remove_dir_all(&temp_dir);
-        std::fs::create_dir_all(&temp_dir).unwrap();
-        let f1 = temp_dir.join("1.md");
-        let f2 = temp_dir.join("2.md");
-        std::fs::write(&f1, "# First").unwrap();
-        std::fs::write(&f2, "# Second").unwrap();
-
-        app.process_action(
-            &egui::Context::default(),
-            AppAction::OpenMultipleDocuments(vec![f1.clone(), f2.clone()]),
-        );
-
-        // Simulate frame updates clearing the background pending document queue
-        while let Some(path) = app.pending_document_loads.pop_front() {
-            app.handle_select_document(path, false);
-        }
-
-        // Both documents are opened
-        assert_eq!(app.state.document.open_documents.len(), 2);
-        // First document is activated (loaded) and second stays lazy
-        assert!(app.state.document.open_documents[0].is_loaded);
-        assert!(!app.state.document.open_documents[1].is_loaded);
-        // Active index points to the first document
-        assert_eq!(app.state.document.active_doc_idx, Some(0));
-
-        let _ = std::fs::remove_dir_all(&temp_dir);
-    }
-
-    // Removed redundant AiProviderRegistry and PluginRegistry imports
-    use crate::app_state::AppState;
-    use crate::preview_pane::PreviewPane;
-    use katana_platform::FilesystemService;
-
-    fn setup_test_app() -> KatanaApp {
-        let state = AppState::new(
-            AiProviderRegistry::new(),
-            PluginRegistry::new(),
-            katana_platform::SettingsService::default(),
-            std::sync::Arc::new(katana_platform::InMemoryCacheService::default()),
-        );
-        KatanaApp {
-            state,
-            fs: FilesystemService::new(),
-            pending_action: AppAction::None,
-            tab_previews: Vec::new(),
-            download_rx: None,
-            workspace_rx: None,
-            update_rx: None,
-            changelog_rx: None,
-            update_install_rx: None,
-            export_tasks: Vec::new(),
-            pending_document_loads: std::collections::VecDeque::new(),
-            show_about: false,
-            show_update_dialog: false,
-            update_markdown_cache: egui_commonmark::CommonMarkCache::default(),
-            update_notified: false,
-            about_icon: None,
-            cached_theme: None,
-            cached_font_size: None,
-            cached_font_family: None,
-            settings_preview: PreviewPane::default(),
-            needs_splash: false,
-            splash_start: None,
-            show_meta_info_for: None,
-            pending_relaunch: None,
-            changelog_sections: Vec::new(),
-            needs_changelog_display: false,
-            old_app_version: None,
-        }
-    }
-
-    #[test]
-    fn test_toggle_about_action() {
-        let mut app = setup_test_app();
-        assert!(!app.show_about);
-
-        app.process_action(&egui::Context::default(), AppAction::ToggleAbout);
-        assert!(app.show_about);
-
-        app.process_action(&egui::Context::default(), AppAction::ToggleAbout);
-        assert!(!app.show_about);
-    }
-
-    #[test]
-    fn test_check_for_updates_manual_trigger() {
-        let mut app = setup_test_app();
-        app.state.update.checking = true;
-        // manually trigger again while already checking
-        app.start_update_check(true);
-        // should have skipped spawning another one but set dialog=true
-        assert!(app.show_update_dialog);
-    }
-
-    #[test]
-    fn test_check_for_updates_action() {
-        let mut app = setup_test_app();
-        assert!(!app.show_update_dialog);
-        assert!(!app.state.update.checking);
-
-        // trigger manual update check
-        app.process_action(&egui::Context::default(), AppAction::CheckForUpdates);
-
-        // it should immediately set show_update_dialog = true for manual checks
-        assert!(app.show_update_dialog);
-        assert!(app.state.update.checking);
-
-        // Emulate an update channel response
-        let (tx, rx) = std::sync::mpsc::channel();
-        tx.send(Ok(Some(katana_core::update::ReleaseInfo {
-            tag_name: "100.0.0".to_string(),
-            html_url: "".to_string(),
-            download_url: "".to_string(),
-            body: "".to_string(),
-        })))
-        .unwrap();
-        app.update_rx = Some(rx);
-
-        let ctx = eframe::egui::Context::default();
-        app.poll_update_check(&ctx);
-
-        assert!(app.state.update.available.is_some());
-        assert_eq!(
-            app.state.update.available.as_ref().unwrap().tag_name,
-            "100.0.0"
-        );
-        assert!(app.update_rx.is_none());
-        assert!(app.update_notified);
-    }
-
-    #[test]
-    fn test_update_check_error_action() {
-        let mut app = setup_test_app();
-        app.state.update.checking = true;
-        let (tx, rx) = std::sync::mpsc::channel();
-        tx.send(Err("Network failure".to_string())).unwrap();
-        app.update_rx = Some(rx);
-
-        let ctx = eframe::egui::Context::default();
-        app.poll_update_check(&ctx);
-
-        assert_eq!(app.state.update.check_error.unwrap(), "Network failure");
-        assert!(app.update_rx.is_none());
-    }
-
-    #[test]
-    fn test_update_check_channel_closed() {
-        let mut app = setup_test_app();
-        app.state.update.checking = true;
-        let (tx, rx) =
-            std::sync::mpsc::channel::<Result<Option<katana_core::update::ReleaseInfo>, String>>();
-        drop(tx); // cause Err(RecvError) or Disconnected
-        app.update_rx = Some(rx);
-
-        let ctx = eframe::egui::Context::default();
-        app.poll_update_check(&ctx);
-
-        assert!(!app.state.update.checking);
-        assert!(app.update_rx.is_none());
-    }
-
-    #[test]
-    fn test_background_update_check_shows_dialog_only_once() {
-        let mut app = setup_test_app();
-        app.start_update_check(false); // background check
-        assert!(!app.show_update_dialog); // should be hidden during check
-        assert!(!app.update_notified);
-
-        let (tx, rx) = std::sync::mpsc::channel();
-        tx.send(Ok(Some(katana_core::update::ReleaseInfo {
-            tag_name: "100.0.0".to_string(),
-            html_url: "".to_string(),
-            download_url: "".to_string(),
-            body: "".to_string(),
-        })))
-        .unwrap();
-        app.update_rx = Some(rx);
-
-        let ctx = eframe::egui::Context::default();
-        app.poll_update_check(&ctx);
-
-        // Now since it's newer and we weren't notified, it should pop up
-        assert!(app.show_update_dialog);
-        assert!(app.update_notified);
-    }
-
-    // ── export_html_to_tmp tests ──
-
-    #[test]
-    pub(crate) fn export_html_to_tmp_writes_html_file() {
-        let preset = katana_core::markdown::color_preset::DiagramColorPreset::dark();
-        let filename = "katana_test_export.html";
-        let result = super::export_html_to_tmp("# Hello", filename, preset, None);
-        let path = result.unwrap();
-        assert!(path.exists(), "HTML file must exist at {}", path.display());
-        let contents = std::fs::read_to_string(&path).unwrap();
-        assert!(
-            contents.contains("<!DOCTYPE html>"),
-            "Output must be valid HTML"
-        );
-        assert!(contents.contains("Hello"), "Output must contain heading");
-        // Cleanup
-        let _ = std::fs::remove_file(&path);
-    }
-
-    #[test]
-    pub(crate) fn export_html_to_tmp_path_is_in_temp_dir() {
-        let preset = katana_core::markdown::color_preset::DiagramColorPreset::dark();
-        let filename = "katana_path_check.html";
-        let path = super::export_html_to_tmp("test", filename, preset, None).unwrap();
-        let expected = std::path::PathBuf::from("/tmp").join(filename);
-        assert_eq!(path, expected);
-        let _ = std::fs::remove_file(&path);
-    }
-
-    // ── export_as_html integration tests ──
-
-    #[test]
-    fn export_as_html_creates_task_with_open_on_complete() {
-        let mut app = make_app();
-        let dir = tempfile::tempdir().unwrap();
-        let md_path = dir.path().join("hello.md");
-        std::fs::write(&md_path, "# Integration Test").unwrap();
-        app.handle_select_document(md_path.clone(), true);
-
-        app.export_as_html(&egui::Context::default(), "# Integration Test", &md_path);
-
-        assert_eq!(
-            app.export_tasks.len(),
-            1,
-            "must push exactly one ExportTask"
-        );
-        let task = &app.export_tasks[0];
-        assert!(
-            task.open_on_complete,
-            "HTML export must set open_on_complete = true"
-        );
-        assert!(
-            task.filename.ends_with(".html"),
-            "filename must be .html, got {}",
-            task.filename
-        );
-    }
-
-    #[test]
-    fn export_as_html_thread_produces_html_file_in_tmp() {
-        let mut app = make_app();
-        let dir = tempfile::tempdir().unwrap();
-        let md_path = dir.path().join("real_doc.md");
-        std::fs::write(&md_path, "# Real Document\n\nParagraph content.").unwrap();
-        app.handle_select_document(md_path.clone(), true);
-
-        app.export_as_html(
-            &egui::Context::default(),
-            "# Real Document\n\nParagraph content.",
-            &md_path,
-        );
-
-        // Wait for the background thread to finish (max 5s).
-        let task = &app.export_tasks[0];
-        let result = task.rx.recv_timeout(std::time::Duration::from_secs(5));
-        let path = result
-            .expect("channel must receive within 5s")
-            .expect("export must succeed");
-
-        // Verify the file is in /tmp
-        assert!(
-            path.starts_with("/tmp"),
-            "path must be under /tmp, got {}",
-            path.display()
-        );
-        assert!(path.exists(), "HTML file must exist at {}", path.display());
-
-        // Verify content is valid HTML
-        let html = std::fs::read_to_string(&path).unwrap();
-        assert!(
-            html.contains("<!DOCTYPE html>"),
-            "must be full HTML document"
-        );
-        assert!(html.contains("Real Document"), "must contain the heading");
-        assert!(html.contains("Paragraph content"), "must contain body text");
-
-        // Cleanup
-        let _ = std::fs::remove_file(&path);
-    }
-
-    #[test]
-    fn export_as_html_multiple_calls_create_multiple_tasks() {
-        let mut app = make_app();
-        let dir = tempfile::tempdir().unwrap();
-        let path1 = dir.path().join("doc1.md");
-        let path2 = dir.path().join("doc2.md");
-        std::fs::write(&path1, "# Doc 1").unwrap();
-        std::fs::write(&path2, "# Doc 2").unwrap();
-        app.handle_select_document(path1.clone(), true);
-
-        app.export_as_html(&egui::Context::default(), "# Doc 1", &path1);
-        app.export_as_html(&egui::Context::default(), "# Doc 2", &path2);
-
-        assert_eq!(
-            app.export_tasks.len(),
-            2,
-            "two exports must create two tasks"
-        );
-
-        // Both complete successfully
-        for task in &app.export_tasks {
-            let result = task.rx.recv_timeout(std::time::Duration::from_secs(5));
-            let path = result.unwrap().unwrap();
-            assert!(path.exists());
-            let _ = std::fs::remove_file(&path);
-        }
-    }
-
-    // ── RED: HTML export bug detection tests ──
-    // Failure mode 1: HTML generation itself fails
-    // Failure mode 2: Path cannot be resolved / opened
-
-    #[test]
-    pub(crate) fn export_html_to_tmp_path_is_canonicalizable() {
-        // If this fails, the generated path has unresolvable symlinks / broken components.
-        let preset = katana_core::markdown::color_preset::DiagramColorPreset::dark();
-        let path = super::export_html_to_tmp("# Test", "katana_canon_test.html", preset, None)
-            .expect("generation must succeed");
-        let canonical = path
-            .canonicalize()
-            .unwrap_or_else(|e| panic!("path {} must be canonicalizable: {e}", path.display()));
-        assert!(
-            canonical.is_absolute(),
-            "canonical path must be absolute: {}",
-            canonical.display()
-        );
-        let _ = std::fs::remove_file(&path);
-    }
-
-    #[test]
-    fn export_html_file_url_is_valid_and_openable() {
-        // Reproduces the exact URL construction from poll_export.
-        let preset = katana_core::markdown::color_preset::DiagramColorPreset::dark();
-        let path = super::export_html_to_tmp("# URL Test", "katana_url_test.html", preset, None)
-            .expect("generation must succeed");
-
-        // Construct URL exactly as poll_export does (line ~1035)
-        let url = format!("file://{}", path.display());
-
-        // file:// + absolute path (/tmp/...) = file:///tmp/... — must have 3 slashes
-        assert!(
-            url.starts_with("file:///"),
-            "URL must start with file:/// (3 slashes), got: {url}"
-        );
-
-        // Extract path from URL and verify it exists
-        let file_path = std::path::Path::new(url.strip_prefix("file://").unwrap());
-        assert!(
-            file_path.exists(),
-            "path extracted from URL must exist: {} (url={url})",
-            file_path.display()
-        );
-
-        // Canonicalize to catch symlink issues (macOS /var -> /private/var)
-        let canonical = path.canonicalize().unwrap();
-        let canonical_url = format!("file://{}", canonical.display());
-        let canonical_file_path =
-            std::path::Path::new(canonical_url.strip_prefix("file://").unwrap());
-        assert!(
-            canonical_file_path.exists(),
-            "canonical path from URL must exist: {}",
-            canonical_file_path.display()
-        );
-
-        let _ = std::fs::remove_file(&path);
-    }
-}
+include!("shell_tests.rs");
